exam_backend/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Authentication
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9

# Cache & Async
redis>=5.0.0
django-redis>=5.4.0
celery>=5.3.0

# API Documentation
drf-spectacular>=0.27.0

# Filtering & Search
django-filter>=23.5
drf-nested-routers>=0.93.4

# Environment
python-dotenv>=1.0.0
python-decouple>=3.8

# File Storage
django-storages>=1.14.0
boto3>=1.34.0

# Utils
Pillow>=10.0.0

# Development
django-debug-toolbar>=4.2.0

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
factory-boy>=3.3.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0
